let user create new documents
Added a form for document creation, for now any role can create a new
document. Later on this might be limited to certain roles, like admin
only for example.
In EditorController the POST mapping receiving the form is afterwards
redirecting back to the GET mapping, in order to prevent duplicate form
submission.