let users save changes to documents
A save button will appear under an open document, onclick the document
will be saved. However, concurrent editing is not supported yet. Whoever
saves last overrides any work done between their opening the document
and their saving it.

Also changed mapping for creation of a new document, to stay in line
with the "/save" mapping. Additionally after creation of a new document,
users are redirected to the document in order to let them start editing
it right away.