add invitation generator and refactor
Introduced InvitationGeneratorService and moved
InvitationsValidatorService to Invitations package, so they both reside
in the same place.
The invitation generator is using a cryptographically secure RNG to
create new invitation codes, which is meant to make guessing next
invitation code (for example based on the attacker received
legitimately) and creating multiple accounts, impersonating actual
recipients of the codes.
It also takes care of saving the newly generated invitation to the
database.